refactor(mb-api): extraire getFonctionAgregationActive dans indicateur/resolveAgregation
La fonction était dupliquée entre resolveSerieIndividu et resolveObjectifIndividu.
Déplacée dans un fichier partagé avec une signature structurelle minimale (pas de
BaseResolveContext intermédiaire).
